Remove parens around cast
Currently corrode generates `expr as (type)`, but the parenthesis around
the type appear to be unnecessary. There are only two cases where they
are present in the rust code base, one is in a precedence test to check
that `3 as (usize) * 3 == 3 as usize * 3`, and in the other they can be
removed.